chore: storybook - styled component의 전역 변수 설정
