--- .storybook/preview.ts
@@ -1,16 +1,0 @@
-import type { Preview } from "@storybook/react";
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-    layout: "centered",
-    actions: { argTypesRegex: "^on.*" },
-  },
-};
-
-export default preview;
